apps/backend/requirements.txt
fastapi[standard]==0.115.11
sqlalchemy[asyncio]==2.0.38
alembic==1.14.1
pydantic[email]==2.10.6
pydantic-settings==2.8.1
psycopg[binary]==3.2.5
asyncpg==0.30.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
influxdb-client>=1.48.0
python-multipart==0.0.20
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.61.1
numpy==2.2.3

pytz==2024.1